refactor(core): deprecated core module in favour of root responsibility
