+++ src/components/Header/parts/SignIn/index.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import type { FC } from 'react'
 
-import { Routes } from '~/features/core/constants/routes'
+import { Routes } from '~/constants/routes'
 
 import { Anchor } from './styles'
