--- src/features/core/index.ts
@@ -1,3 +1,0 @@
-export { Routes } from './constants/routes'
-export { HeadDefault } from './components/HeadDefault'
-export { ErrorBoundary } from './components/ErrorBoundary'
